+++ packages/leancode_cubit_utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * **BREAKING CHANGE**: Remove `map` and `isEmpty` methods from `RequestCubit`.
 * **BREAKING CHANGE**: Remove `TData` generic parameter from `RequestCubit`.
+* Add `map` method to `RequestState`.
+* Add `onRefreshing` callback to `RequestCubitBuilder`.
+* **BREAKING CHANGE**: Make `data` in `RequestRefreshState` non-nullable.
+* **BREAKING CHANGE**: Add `data` to `RequestEmptyState`.
 
 ## 0.4.2